Add worker thread to check hex -> callsign
Fairly commonly the callsign that we get from VDL2 messages doesn't
actually match the 'canonical' callsign for a flight. A simple way to
address this problem is to ask an aggregator with good coverage if they
have current data on that hex and extract the 'correct' callsign from
that.

adsb.fi has fairly good coverage and allowed me to use their API for
this purpose, as long as I stay far away from their rate limit.
For now I call the API no more than once every five seconds.
In order to make this work we send all the requests to check routes to a
worker thread where we combine the hex'es and look them up in a combined
API call.

The code also tries to cache recent lookups and to reduce the overall
checks and calculations that it makes. The timeouts for these cache
values are of course debatable.

It would be good to add other API backends besides adsb.fi to get
broader coverage of the flights that we might be getting ACARS data for.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>